+++ BlueprintUILists/Sources/List.swift
@@ -50,8 +50,8 @@ public struct List : Element
     
     /// How the `List` is measured when the element is laid out
     /// by Blueprint.  Defaults to `.fillParent`, which means
-    /// it will take up all the size it is given. You can change this to
-    /// `.measureContent` to instead measure the optimal size.
+    /// it will take up all the height it is given. You can change this to
+    /// `.measureContent` to instead measure the optimal height.
     ///
     /// See the `List.Measurement` documentation for more.
     public var measurement : List.Measurement
@@ -122,19 +122,25 @@ extension List {
             properties.environment.blueprintEnvironment = environment
             
             self.properties = properties
-            self.measurement = measurement
+            
+            if measurement.needsMeasurement {
+                self.measurement = measurement
+            } else {
+                self.measurement = .fillParent
+            }
         }
         
         // MARK: Element
             
         public var content : ElementContent {
+            
             switch self.measurement {
             case .fillParent:
                 return ElementContent { constraint -> CGSize in
                     constraint.maximum
                 }
                 
-            case .measureContent(let key, let limit):
+            case .measureContent(let key, let horizontalFill, let verticalFill, let limit):
                 return ElementContent(
                     measurementCachingKey: {
                         if let key = key {
@@ -144,11 +150,32 @@ extension List {
                         }
                     }()
                 ) { constraint -> CGSize in
-                    ListView.contentSize(
+                    
+                    let size = ListView.contentSize(
                         in: constraint.maximum,
                         for: self.properties,
                         itemLimit: limit
                     )
+                    
+                    let width : CGFloat = {
+                        switch horizontalFill {
+                        case .fillParent:
+                            return constraint.maximum.width
+                        case .natural:
+                            return size.naturalWidth ?? size.contentSize.width
+                        }
+                    }()
+                    
+                    let height : CGFloat = {
+                        switch verticalFill {
+                        case .fillParent:
+                            return constraint.maximum.width
+                        case .natural:
+                            return size.contentSize.height
+                        }
+                    }()
+                    
+                    return CGSize(width: width, height: height)
                 }
             }
         }