+++ ListableUI/Sources/Layout/ListLayout/ListLayoutContent.swift
@@ -13,6 +13,8 @@ public final class ListLayoutContent
 {
     public var contentSize : CGSize
     
+    public var naturalContentWidth : CGFloat?
+    
     public let containerHeader : SupplementaryItemInfo
     public let header : SupplementaryItemInfo
     public let footer : SupplementaryItemInfo
@@ -37,9 +39,16 @@ public final class ListLayoutContent
         return all
     }
     
+    public func maxMeasuredValue(for keyPath : KeyPath<CGSize, CGFloat>) -> CGFloat {
+        self.all.reduce(0) { value, item in
+            max(value, item.measuredSize[keyPath: keyPath])
+        }
+    }
+    
     init()
     {
         self.contentSize = .zero
+        self.naturalContentWidth = nil
         
         self.containerHeader = .empty(.listContainerHeader)
         self.header = .empty(.listHeader)
@@ -57,6 +66,7 @@ public final class ListLayoutContent
         sections : [SectionInfo]
     ) {
         self.contentSize = .zero
+        self.naturalContentWidth = nil
         
         self.containerHeader = containerHeader ?? .empty(.listContainerHeader)
         self.header = header ?? .empty(.listHeader)
@@ -230,6 +240,8 @@ public final class ListLayoutContent
     var layoutAttributes : ListLayoutAttributes {
         ListLayoutAttributes(
             contentSize: self.contentSize,
+            naturalContentWidth : self.naturalContentWidth,
+            containerHeader: self.containerHeader.isPopulated ? .init(frame: self.containerHeader.defaultFrame) : nil,
             header: self.header.isPopulated ? .init(frame: self.header.defaultFrame) : nil,
             footer: self.footer.isPopulated ? .init(frame: self.footer.defaultFrame) : nil,
             overscrollFooter: self.overscrollFooter.isPopulated ? .init(frame: self.overscrollFooter.defaultFrame) : nil,
@@ -250,6 +262,8 @@ public final class ListLayoutContent
 
 public protocol ListLayoutContentItem : AnyObject
 {
+    var measuredSize : CGSize { get set }
+    
     var size : CGSize { get set }
     var x : CGFloat { get set }
     var y : CGFloat { get set }
@@ -343,6 +357,8 @@ extension ListLayoutContent
                 
         public let isPopulated : Bool
                         
+        public var measuredSize : CGSize = .zero
+        
         public var size : CGSize = .zero
         
         public var x : CGFloat = .zero
@@ -410,6 +426,8 @@ extension ListLayoutContent
         
         public var position : ItemPosition = .single
                 
+        public var measuredSize : CGSize = .zero
+        
         public var size : CGSize = .zero
                 
         public var x : CGFloat = .zero