+++ ListableUI/Sources/Layout/Table/TableListLayout.swift
@@ -478,6 +478,8 @@ final class TableListLayout : ListLayout
 
         let size = headerFooter.measurer(measureInfo)
         
+        headerFooter.measuredSize = size
+        
         // Write the measurement and position out to the header/footer.
         
         self.direction.switch(
@@ -538,6 +540,10 @@ final class TableListLayout : ListLayout
         self.setItemPositions()
         
         delegate?.listViewLayoutUpdatedItemPositions()
+        
+        //
+        // Sizing
+        //
         
         var contentBottom : CGFloat = 0.0
                 
@@ -649,6 +655,8 @@ final class TableListLayout : ListLayout
                     
                     let size = item.measurer(measureInfo)
                     
+                    item.measuredSize = size
+                    
                     self.direction.switch(
                         vertical: {
                             item.x = itemPosition.origin
@@ -705,6 +713,8 @@ final class TableListLayout : ListLayout
                         )
                                                 
                         let size = item.measurer(measureInfo)
+                        
+                        item.measuredSize = size
                         
                         let height = self.direction.switch(vertical: size.height, horizontal: size.width)
                         
@@ -812,6 +822,12 @@ final class TableListLayout : ListLayout
         //
         
         self.content.contentSize = self.direction.size(for: CGSize(width: viewWidth, height: contentBottom))
+        
+        self.content.naturalContentWidth = self.direction.switch {
+            return content.maxMeasuredValue(for: \.width) + bounds.padding.right + bounds.padding.left
+        } horizontal: {
+            return content.maxMeasuredValue(for: \.width) + bounds.padding.top + bounds.padding.bottom
+        }
     }
     
     private func setItemPositions()